Unit test for SaxWalkerOverDom.
I found some kind of bug in the Java SAX and DOM parsers.  If the last
element only contains an attribute, both the DOM and SAX parsers give
that element the URI, local name, and QName of the root element
("schema").  The SAX parser will give it the correct attributes, but the
DOM parser will give it the root element's attributes as well, causing
the TestSaxWalkerOverDom test to fail.

To work around this, I added a dummy element at the end.