Adding support for wildcard elements.
At this time, it is not possible to determine the XmlSchemaAny's
original target namespace without knowing the original element that
owned it.  Likewise, unless the XmlSchemaAny's namespace is an actual
namespace, or ##any, or ##local, there is no way to validate it.

The work-around is to walk upwards through the tree and find the owning
element, then use its namespace as the target namespace.