Tests don't pass yet, but committing here.
We need to keep track of where we are in the children.  This way when we
don't need to backtrack our steps when we start a new element under an
existing one.

Next up: every time we increase the number of occurrences, we need a new
instance of that node in the tree.