Flushed out the path manager, including following & unfollowing paths.
Not sure if I committed my work to build a tree of possiblities, but it
turned out to not be feasible.  I only have access to the element names
when trying to follow new possiblities, so all possible paths must be
rooted at an element.  Likewise, I cannot create branches at arbitrary
locations in the path.