Bug Fixes:
(1) Do not add prefixes to attributes in the root element.  This is
apparently invalid.
(2) If working with an unsigned long, and the read double value is
larger than the max, use the max instead.
(3) Preserve empty strings when possible.
(4) Run the test on all three XML documents.
(5) Modify the test itself to use values that convert well to decimal
types.  Unfortunately, not all BigDecimal values can be easily
represented.