Painted myself into a corner.  In order to properly determine if an attribute should get a prefix, we need to know if it is in the global namespace or not.
The problem is an attributes both in the global namespace, and outside
of the global namespace, can co-exist with the same name.  The only way
to determine which is which is to find the attributes tied to the
element you're looking at, and see if it is top-level or not.

However, this information is long gone by the time we reach
DomBuilderFromSax.  Attributes are copied to prevent direct
modification, and the copy destroys any information about whether the
attribute is top-level or not.  I don't know how to fix it.  For now,
I'm changing the test to avoid the name conflict.

In other news, SaxWalkerOverDom is surprisingly flimsy ... apparently
even small amounts of content may be spread over multiple nodes, and my
rename of "id" to "mapId" created a content break.  I'm not sure what to
do about that either.

Test Coverage: 75.8%